+++ src/Calculation/Arithmetic/Arithmetic.java
@@ -1,6 +1,6 @@
 package Calculation.Arithmetic;
 
 public interface Arithmetic {
-    public int resultOperation();
+    public double resultOperation();
 
 }
